+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 
-<!-- ## [Unreleased] -->
+## [Unreleased]
+
+### Added
+
+- Support for Lualine theme.
 
 ## [1.4.1] - 2024-09-26
 
@@ -113,7 +117,7 @@
 
 Initial version from [WittyJudge](https://github.com/WIttyJudge).
 
-[unreleased]: https://github.com/f4z3r/gruvbox-material.nvim/compare/v1.4.0...HEAD
+[unreleased]: https://github.com/f4z3r/gruvbox-material.nvim/compare/v1.4.1...HEAD
 [1.4.1]: https://github.com/f4z3r/gruvbox-material.nvim/compare/v1.4.0...v1.4.1
 [1.4.0]: https://github.com/f4z3r/gruvbox-material.nvim/compare/v1.3.1...v1.4.0
 [1.3.1]: https://github.com/f4z3r/gruvbox-material.nvim/compare/v1.3.0...v1.3.1